+++ MoreShipUpgrades/Misc/AssetBundleHandler.cs
@@ -148,6 +148,7 @@ namespace MoreShipUpgrades.Misc
             { IgnitionCoil.UPGRADE_NAME, upgrades+"IgnitionCoil.prefab" },
             { WeedGeneticManipulation.UPGRADE_NAME, upgrades+"WeedGeneticManipulation.prefab" },
             { FedoraSuit.UPGRADE_NAME, upgrades+"FedoraSuit.prefab" },
+            { TurboTank.UPGRADE_NAME, upgrades+"TurboTank.prefab" },
 
             { "Advanced Portable Tele", storeItems+"TpButtonAdv.asset" },
             { "Portable Tele", storeItems+"TpButton.asset" },